+++ RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+### 0.1.2  - 2020-04-15
+* Fix #5: show moves / capacity
+* Collapsible players dashboard
+
 ### 0.1.1 - 2020-04-15
 * Async server